backend/requirements.txt
Django>=2.0,<3.0
djangorestframework
django-cors-headers
django-rest-knox
